chore: move all constants to a separate file and refactor
